Use MediaSessionCompat for the headset button
