+++ handheld/src/main/AndroidManifest.xml
@@ -142,11 +142,16 @@
         </activity>
 
         <!-- Media button -->
-        <receiver android:name="org.jraf.android.bikey.app.mediabutton.MediaButtonReceiver">
+        <receiver android:name="android.support.v4.media.session.MediaButtonReceiver" >
             <intent-filter>
                 <action android:name="android.intent.action.MEDIA_BUTTON" />
             </intent-filter>
         </receiver>
+        <service android:name="org.jraf.android.bikey.app.mediabutton.MediaButtonService">
+            <intent-filter>
+                <action android:name="android.intent.action.MEDIA_BUTTON" />
+            </intent-filter>
+        </service>
 
         <!-- Heart rate monitor scan -->
         <activity
